requirements.txt
streamlit
google-genai
python-dotenv
langchain
faiss-cpu
langchain_google_genai